docvault-rag/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic-settings==2.3.0
httpx>=0.27.0

# PDF extraction
pymupdf>=1.24.0

# Text chunking
langchain-text-splitters>=0.2.0

# Gemini embeddings
google-generativeai>=0.7.0

# Vector store
chromadb>=0.5.0
